# docs/channel/volc-passthrough/apifox-seedance2.openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "seedance-2官方透传",
    "version": "1.0.0",
...
    "x-apifox-folder": "seedance-2官方透传"
  },
  "servers": [
    {
      "url": "{{baseUrl}}",
      "description": "站点地址。在 Apifox 环境变量里把 baseUrl 设为 https://your-site.com（不带尾斜杠）"
    }
  ],
...
  "paths": {
...
    "/volc/?Action=ListAssets&Version=2024-01-01": {
      "post": {
        "tags": [
          "二、控制面 · 12 个 Action（逐个）"
        ],
        "summary": "3.2 ListAssets 素材列表",
        "operationId": "ListAssets",
        "description": "`Filter` 必选，**`Filter.GroupType` 必选**。\n\n| 字段 | 类型 | 说明 |\n|---|---|---|\n| `Filter.GroupType` | string | **必选**。`AIGC` ｜ `LivenessFace` |\n| `Filter.GroupIds` | string[] | 限定素材组 |\n| `Filter.Name` | string | 名称**模糊搜索**，≤64 字符 |\n| `Filter.Statuses` | string[] | `Active` ｜ `Processing` ｜ `Failed` |\n| `NextToken` / `MaxResults` | string / int | 游标分页，[1,100]，默认 10 |\n| `PageNumber` / `PageSize` | int / int | 页码分页（兼容写法） |\n| `SortBy` | string | `CreateTime`（默认）｜ `UpdateTime` ｜ `GroupId` |\n| `SortOrder` | string | `Desc`（默认）｜ `Asc` |\n\n分页规则同 2.2：两种方式互斥；游标分页**不返回** `TotalCount` / `PageNumber` / `PageSize`，**`NextToken` 缺失即遍历结束**；翻页时 `Filter` / `SortBy` / `SortOrder` 应与首次查询保持一致。深翻页阈值这里是 `PageSize * PageNumber > 20000`（素材组是 10000，两者不同）。\n\n**响应已按归属过滤，只含你自己的素材。**",
        "security": [
          {
            "siteToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "examples": {
                "列出可用素材": {
                  "summary": "最常用。只列 Active 的，Processing 的不能用",
                  "value": "{\n  \"Filter\": {\n    \"GroupType\": \"AIGC\",\n    \"Statuses\": [\n      \"Active\"\n    ]\n  },\n  \"MaxResults\": 20,\n  \"SortBy\": \"CreateTime\",\n  \"SortOrder\": \"Desc\"\n}"
                },
                "列出某组下的素材": {
                  "summary": "按组筛选",
                  "value": "{\n  \"Filter\": {\n    \"GroupType\": \"AIGC\",\n    \"GroupIds\": [\n      \"assetgroup-20260410113901-qk2mv\"\n    ]\n  },\n  \"MaxResults\": 50\n}"
                },
                "按状态列出处理中的素材": {
                  "summary": "巡检用：看有没有卡在 Processing 或失败的",
                  "value": "{\n  \"Filter\": {\n    \"GroupType\": \"AIGC\",\n    \"Statuses\": [\n      \"Processing\",\n      \"Failed\"\n    ]\n  },\n  \"MaxResults\": 20\n}"
                },
                "按名称模糊搜索": {
                  "summary": "Name 是模糊匹配，与 2.2 的精确匹配不同",
                  "value": "{\n  \"Filter\": {\n    \"GroupType\": \"AIGC\",\n    \"Name\": \"主角\"\n  }\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功。HTTP 200，信封为 {ResponseMetadata, Result}",
            "content": {
              "application/json": {
                "examples": {
                  "成功": {
                    "summary": "游标分页不返回 TotalCount —— 这是正常的，不是漏字段",
                    "value": "{\n  \"ResponseMetadata\": {\n    \"RequestId\": \"202609140804468423AF90FD6CA6BA26F6\",\n    \"Action\": \"ListAssets\",\n    \"Version\": \"2024-01-01\",\n    \"Service\": \"ark\",\n    \"Region\": \"cn-beijing\"\n  },\n  \"Result\": {\n    \"Items\": [\n      {\n        \"Id\": \"asset-20260410114236-8cdfz\",\n        \"Name\": \"主角正脸\",\n        \"AssetType\": \"Image\",\n        \"GroupId\": \"assetgroup-20260410113901-qk2mv\",\n        \"Status\": \"Active\",\n        \"ProjectName\": \"default\",\n        \"CreateTime\": \"2026-04-10T11:42:36Z\",\n        \"UpdateTime\": \"2026-04-10T11:42:41Z\"\n      }\n    ],\n    \"NextToken\": \"\"\n  }\n}"
                  },
                  "【失败态】素材 Processing / Failed": {
                    "summary": "Failed 的素材不会自动重试，要删掉重新 CreateAsset",
                    "value": "{\n  \"ResponseMetadata\": {\n    \"RequestId\": \"202609140804468423AF90FD6CA6BA26F6\",\n    \"Action\": \"ListAssets\",\n    \"Version\": \"2024-01-01\",\n    \"Service\": \"ark\",\n    \"Region\": \"cn-beijing\"\n  },\n  \"Result\": {\n    \"Items\": [\n      {\n        \"Id\": \"asset-20260410115000-proc1\",\n        \"Name\": \"主角正脸-v2\",\n        \"AssetType\": \"Image\",\n        \"GroupId\": \"assetgroup-20260410113901-qk2mv\",\n        \"Status\": \"Processing\"\n      },\n      {\n        \"Id\": \"asset-20260410115100-fail1\",\n        \"Name\": \"配角正脸\",\n        \"AssetType\": \"Image\",\n        \"GroupId\": \"assetgroup-20260410113901-qk2mv\",\n        \"Status\": \"Failed\",\n        \"Error\": {\n          \"Code\": \"ContentRestricted\",\n          \"Message\": \"input image sensitive content detected\"\n        }\n      }\n    ],\n    \"NextToken\": \"\"\n  }\n}"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "siteToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "站点令牌，形如 sk-xxxxxxxx。数据面与控制面透传都用它：Authorization: Bearer {{token}}。在 Apifox 环境变量里配置 token，并把本方案设为全局认证。真值请填自己站点的令牌，不要用示例里的占位符。"
      },
      "siteCookie": {
        "type": "apiKey",
        "in": "header",
        "name": "Cookie",
        "description": "站点登录态，形如 session=xxxxx。仅 /api/volc_asset/* 这五个站点内部接口用；也可改用站点 Access Token 走同名头部。"
      }
    },
...
  },
  "security": [
    {
      "siteToken": []
    }
  ]
}